Move ByteGenerator to separate file
